Commited a few small changes
