+++ wdio/screen-objects/testIDs/Screens/TransactionConfirm.testIds.js
@@ -1,0 +1,2 @@
+// eslint-disable-next-line import/prefer-default-export
+export const COMFIRM_TXN_AMOUNT = 'confirm-txn-amount';
